Add files via upload
This project facilitates a student login system where individuals can enter their Registration Number and Password into a secure login form. Upon submission, the program authenticates the user's credentials. If verified, the system displays the student's Registration Number and Name. 

In instances where a student is not registered, they are prompted to complete a Registration form, the details of which are then securely stored in the database. Following this process, the student is redirected to the login page to authenticate their account.

Technologies utilized: MySQL Connector (to establish a connection between the Java application and XAMPP for PHPMyAdmin).

Authentication details employed: (RegNum: "H230407R") (Password: "admin").